refactor css architecture to 7-1 and start split components to make them reusable
