+++ src/assets/styles/_pages.scss
@@ -1,11 +1,4 @@
 @import './pages/homePage';
-@import './pages/newbie';
+@import './pages/newbieCardPage';
 
-#app {
-    height: 100%;
-    min-height: 100vh;
-    width: 100%;
-    max-width: 1920px;
-    margin: auto;
-    overflow: hidden;
-}
+@import './components/newbie/newbie';
