+++ src/assets/styles/base/_base.scss
@@ -1,0 +1,8 @@
+html {
+    /* 62.5% of 16px browser font size is 10px */
+    font-size: 62.5%;
+}
+
+body {
+    font-size: 1.6rem;
+}
