+++ src/assets/styles/components/newbie/_interactiveRatingCardView.scss
