+++ src/assets/styles/components/newbie/_newbie.scss
@@ -1,0 +1,10 @@
+@import './qrCodePage';
+@import './productPreviewCard';
+@import './resultSummary';
+@import './3columnPreview';
+@import './interactiveRatingCardView';
+@import './nftPreviewCard';
+@import './orderSummaryCard';
+@import './statsPreviewCard';
+@import './socialLinksProfileCard';
+
