+++ src/assets/styles/components/newbie/_nftPreviewCard.scss
