+++ src/assets/styles/components/newbie/_orderSummaryCard.scss
