+++ src/assets/styles/components/newbie/_productPreviewCard.scss
@@ -1,0 +1,161 @@
+.product-preview-card {
+    @import url('https://fonts.googleapis.com/css2?family=Fraunces:opsz,wght@9..144,700&family=Montserrat:wght@500;700&display=swap');
+
+    $light-text-color: #6C7289;
+    $title-text-color: #1C232B;
+    $green: #3D8168;
+    $dark-green: #1A4031;
+    $border-radius: 1rem;
+    $product-mobile-img: "/assets/images/productPreview/image-product-mobile.jpg";
+    $product-desktop-img: "/assets/images/productPreview/image-product-desktop.jpg";
+
+    width: 100%;
+    max-width: 343px;
+    background-color: white;
+    border-radius: $border-radius;
+    font-family: 'Montserrat', sans-serif;
+
+
+    @media (min-width: 1440px) {
+        margin-left: -236px;
+    }
+
+    .product-preview-card-img {
+        max-width: 100%;
+        height: 240px;
+        border-top-left-radius: $border-radius;
+        border-top-right-radius: $border-radius;
+        background: lightgrey url($product-mobile-img) no-repeat center / cover;
+    }
+
+    .product-preview-card-title,
+    .product-preview-card-price {
+        font-family: 'Fraunces', serif;
+    }
+
+    .product-preview-card-content-wrapper {
+        max-width: 295px;
+        margin: 2.4rem;
+
+        .product-preview-card-tag {
+            color: $light-text-color;
+            text-transform: uppercase;
+            font-weight: 500;
+            font-size: 1.2rem;
+            letter-spacing: 5px;
+            margin-bottom: 1.2rem;
+        }
+
+        .product-preview-card-title {
+            color: $title-text-color;
+            font-weight: 700;
+            font-size: 3.2rem;
+            line-height: 32px;
+            margin-bottom: 1.6rem;
+            
+            @media (max-width: 250px) {
+                font-size: 2.2rem;
+            }
+        }
+
+        .product-preview-card-description {
+            color: $light-text-color;
+            font-size: 1.4rem;
+            line-height: 23px;
+            margin-bottom: 2.4rem;
+        }
+
+        .product-preview-card-price-wrapper {
+            display: flex;
+            align-items: center;
+            flex-wrap: wrap;
+            gap: 1.9rem;
+            margin-bottom: 2rem;
+
+            .product-preview-card-price {
+                color: $green;
+                font-size: 3.2rem;
+                font-weight: 700;
+                line-height: 3.2rem;
+            }
+
+            .product-preview-card-old-price {
+                color: $light-text-color;
+                text-decoration: line-through;
+                font-size: 1.3rem;
+                font-weight: 500;
+                line-height: 23px;
+            }
+        }
+
+        .product-preview-card-button {
+            max-width: 295px;
+            width: 100%;
+            background-color: $green;
+            color: white;
+            border: none;
+            @include center-with-flex;
+            flex-wrap: wrap;
+            gap: 1.2rem;
+            border-radius: 0.8rem;
+            margin: auto;
+            padding: 1.6rem 0;
+            font-size: 1.4rem;
+            font-weight: 700;
+            transition: background-color ease-in-out 250ms;
+
+            &:hover {
+                background-color: $dark-green;
+                cursor: pointer;
+            }
+        }
+    }
+
+    @media (min-width: 1440px) {
+        width: 600px;
+        height: 450px;
+        display: flex;
+        align-items: stretch;
+
+        .product-preview-card-img {
+            min-width: 100%;
+            width: 337px;
+            height: 100%;
+            background-color: white;
+            background-image: url($product-desktop-img);
+            border-bottom-left-radius: $border-radius;
+            border-top-right-radius: 0px;
+        }
+
+        .product-preview-card-content-wrapper {
+            box-sizing: content-box;
+            width: 236px;
+            padding: 3.4rem 3rem;
+            background-color: white;
+            border-radius: $border-radius;
+            border-top-left-radius: 0;
+            border-bottom-left-radius: 0;
+            margin: 0;
+
+            .product-preview-card-tag {
+                margin-bottom: 2rem;
+            }
+
+            .product-preview-card-title {
+                margin-bottom: 2.4rem
+            }
+
+            .product-preview-card-description {
+                margin-bottom: 2.9rem;
+            }
+
+            .product-preview-card-price-wrapper {
+                margin-bottom: 3rem;
+            }
+
+            .product-preview-card-button {
+                width: 236px;
+            }
+        }
+    }
+}
