+++ src/assets/styles/components/newbie/_qrCodePage.scss
@@ -1,0 +1,61 @@
+.qr-code-card {
+    @import url('https://fonts.googleapis.com/css2?family=Outfit:wght@400;700&family=Playfair+Display:ital,wght@1,900&family=Roboto:ital,wght@1,300&display=swap');
+
+    $bg-card-color: #FFFFFF;
+    $title-text-color: #1F314F;
+    $description-text-color: #7D889E;
+
+    max-width: 288px;
+    background-color: $bg-card-color;
+    border-radius: 2rem;
+    padding: 1.6rem;
+    box-shadow: 0px 2.5rem 2.5rem 0px rgba(0, 0, 0, 0.05);
+
+    .qr-code-img-wrapper {
+        max-width: 100%;
+        border-radius: 1rem;
+        margin-bottom: 2.4rem;
+
+        .qr-code-img {
+            width: 100%;
+            height: 100%;
+            border-radius: 1rem;
+        }
+    }
+
+    .qr-code-card-title,
+    .qr-code-card-description {
+        font-family: Outfit, sans-serif;
+        font-style: normal;
+        line-height: normal;
+        text-align: center;
+        font-feature-settings: 'clig' off, 'liga' off;
+    }
+
+    .qr-code-card-title {
+        max-width: 286px;
+        font-size: 2.2rem;
+        font-weight: 700;
+        margin-bottom: 1.6rem;
+        color: $title-text-color;
+    }
+
+    .qr-code-card-description {
+        max-width: 256px;
+        margin: 0 auto 1rem;
+        font-size: 1.5rem;
+        font-weight: 400;
+        letter-spacing: 0.188px;
+        color: $description-text-color;
+    }
+
+    @media (max-width: 250px) {
+        .qr-code-card-title {
+            font-size: 1.8rem
+        }
+
+        .qr-code-card-description {
+            font-size: 1.4rem;
+        }
+    }
+}
