+++ src/assets/styles/components/newbie/_socialLinksProfileCard.scss
