+++ src/assets/styles/components/newbie/_statsPreviewCard.scss
