+++ src/assets/styles/main.scss
@@ -1,0 +1,10 @@
+// BASE
+@import './base/reset';
+@import './base/base';
+
+// UTILITIES
+@import './utils/variables';
+@import './utils/mixins';
+
+// PAGES
+@import './pages';
