--- src/assets/styles/pages/_newbie.scss
@@ -1,9 +1,0 @@
-@import './newbie/qrCodePage';
-@import './newbie/productPreviewCard';
-@import './newbie/resultSummary';
-@import './newbie/3columnPreview';
-@import './newbie/interactiveRatingCardView';
-@import './newbie/nftPreviewCard';
-@import './newbie/orderSummaryCard';
-@import './newbie/statsPreviewCard';
-@import './newbie/socialLinksProfileCard';
