--- src/assets/styles/pages/newbie/_qrCodePage.scss
@@ -1,68 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Outfit:wght@400;700&family=Playfair+Display:ital,wght@1,900&family=Roboto:ital,wght@1,300&display=swap');
-
-@media (max-height: 530px) {
-    #newbieQrCode {
-        padding: 12px;
-    }   
-}
-
-#newbieQrCode {
-    $bg-color: #D5E1EF;
-    $bg-card-color: #FFFFFF;
-    $title-text-color: #1F314F;
-    $description-text-color: #7D889E;
-
-    height: 100%;
-    min-height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: $bg-color;
-
-    .qr-code-card {
-        max-height: 465px;
-        max-width: 288px;
-        background-color: $bg-card-color;
-        border-radius: 20px;
-        padding: 16px;
-        box-shadow: 0px 25px 25px 0px rgba(0, 0, 0, 0.05);
-
-        .qr-code-img-wrapper {
-           max-height: 288px;
-           max-width: 100%; 
-           border-radius: 10px;
-           margin-bottom: 24px;
-
-            .qr-code-img {
-                width: 100%;
-                height: 100%;
-                border-radius: 10px;
-            }
-        }
-        .qr-code-card-title,
-        .qr-code-card-description  {
-            font-family: Outfit, sans-serif;
-            font-style: normal;
-            line-height: normal;
-            text-align: center;
-            font-feature-settings: 'clig' off, 'liga' off;
-        }
-
-        .qr-code-card-title {
-            max-width: 286px;
-            font-size: 22px;
-            font-weight: 700;
-            margin-bottom: 16px;   
-            color: $title-text-color;
-        }
-
-        .qr-code-card-description {
-            max-width: 256px;
-            margin: auto;
-            font-size: 15px;
-            font-weight: 400;
-            letter-spacing: 0.188px;
-            color: $description-text-color;
-        }
-    }
-}
