--- src/assets/styles/styles.scss
@@ -1,7 +1,0 @@
-@import './reset';
-
-// VARIABLES
-@import './variables';
-
-// PAGES
-@import './pages';
