+++ src/assets/styles/utils/_mixins.scss
@@ -1,0 +1,5 @@
+@mixin center-with-flex {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
