+++ src/components/NewbieCards/ResultsSummary/ResultsSummarySkill.vue
