+++ src/main.js
@@ -1,4 +1,4 @@
-import './assets/styles/styles.scss'
+import './assets/styles/main.scss'
 
 import { createApp } from 'vue'
 import App from './App.vue'
