+++ src/views/NewbiePages/NewbieCardPages/NewbieCardPage.vue
@@ -1,0 +1,5 @@
+<template>
+    <div id="newbieCardPage">
+        <slot></slot>
+    </div>
+</template>
