+++ src/views/NewbiePages/NewbieCardPages/NewbieQRCodePage.vue
@@ -1,0 +1,9 @@
+<template>
+    <NewbieCardPage class="qr-card-page">
+        <NewbieQRCode/>
+    </NewbieCardPage>
+</template>
+<script setup>
+import NewbieCardPage from './NewbieCardPage.vue';
+import NewbieQRCode from '../../components/NewbieCards/NewbieQRCode.vue';
+</script>./NewbieCardPages/NewbieCardPage.vue
