+++ src/views/NewbiePages/NewbieCardPages/ProductPreviewCardPage.vue
@@ -1,0 +1,10 @@
+<template>
+    <NewbieCardPage class="product-preview-card-page">
+        <ProductPreviewCard />
+    </NewbieCardPage>
+</template>
+
+<script setup>
+import NewbieCardPage from './NewbieCardPage.vue';
+import ProductPreviewCard from '../../../components/NewbieCards/ProductPreviewCard.vue';
+</script>
