+++ src/views/NewbiePages/NewbieCardPages/ResultSummaryPage/ResultSummaryPage.vue
@@ -1,0 +1,14 @@
+<template>
+    <NewbieCardPage class="results-summary-page">
+        <ResultsSummary 
+            :data="data"
+        />
+    </NewbieCardPage>
+</template>
+
+<script setup>
+import NewbieCardPage from '../NewbieCardPage.vue';
+import ResultsSummary from '../../../../components/NewbieCards/ResultsSummary/ResultsSummary.vue';
+import data from './data.json'
+
+</script>
